Insert FoodEaten with Entry Change data type of FoodEaten.amountInGrams

+++ shared/src/commonMain/kotlin/com/faltenreich/diaguard/food/eaten/FoodEaten.kt
@@ -12,7 +12,7 @@ data class FoodEaten(
     override val id: Long,
     override val createdAt: DateTime,
     override val updatedAt: DateTime,
-    val amountInGrams: Double,
+    val amountInGrams: Long,
     val foodId: Long,
     val entryId: Long,
 ) : DatabaseEntity {
